add modify dialog

+++ client/src/pages/dateControl/ModifyMenu.js
@@ -6,12 +6,19 @@ import Button from '@mui/material/Button';
 import Menu from '@mui/material/Menu';
 import MenuItem from '@mui/material/MenuItem';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
+import ModifyDialog from '../../components/dashboard/dialogs/ModifyDialog';
 
 // import actions
 import { handleDataToEdit , updateDate, deleteDate } from '../../actions/dashboardAct';
 
+// import functions
+import { formattingDate } from '../../utils/utils';
+
 const ModifyMenu = ({ id, data }) => {
   const dispatch = useDispatch();
+  const [dialogOpen, setDialogOpen] = useState(false);
+  const [action, setAction] = useState("");
+  const [textContent, setTextContent] = useState([]);
   const [anchorEl, setAnchorEl] = useState(null);
   const open = Boolean(anchorEl);
 
@@ -24,28 +31,33 @@ const ModifyMenu = ({ id, data }) => {
     switch(myValue){
         case 'edit':
             return dispatch(handleDataToEdit(id))
-        case 'open-close': {
-            let obj = { ...data, status: !data.status}
-            return dispatch(updateDate(obj, id))
+        case 'statusChange': {
+            setAction("statusChange")
+            setTextContent(`Want to ${data.status ? 'CLOSE' : 'OPEN'} date: ${formattingDate(new Date(data.openDate), 'dmmy-time')}?`)
+            return setDialogOpen(true)
         }
         case 'delete':
-            return dispatch(deleteDate(id))
+            setAction("delete")
+            setTextContent(`Want to DELETE date: ${formattingDate(new Date(data.openDate), 'dmmy-time')}?`)
+            return setDialogOpen(true)
         default:
             return console.log('do nothing')
     }
-    // if(myValue==="edit"){
-    // //   dispatch(editingExistingBookDate(data))
-    // console.log("edit", data)
-    // } else if(myValue === "closed"){
-    //     console.log("close")
-    // //   let newObjData = {...data, available: false}
-    // //   dispatch(updateExistingBookDate(newObjData, data._id, false))
-    // } else if(myValue === "delete"){
-    //   dispatch(deleteDate(data))
-    // } else {
-    //   console.log("no action executed")
-    // }
   };
+
+  const handleSubmit = () => {
+    setDialogOpen(false);
+    if(action === 'statusChange'){handleStatusChange()}
+    else if(action === 'delete'){
+      dispatch(deleteDate(id))
+    }
+    else { console.log('do nothing, something wrong')}
+  }
+
+  const handleStatusChange = () => {
+    let obj = { ...data, status: !data.status}
+    return dispatch(updateDate(obj, id))
+  }
 
   return (
     <>
@@ -54,11 +66,17 @@ const ModifyMenu = ({ id, data }) => {
         >
         <MoreVertIcon />
         </Button>
-        <Menu id="basic-menu" anchorEl={anchorEl} open={open} onClose={handleClose} MenuListProps={{ 'aria-labelledby': 'basic-button', }}>
-            <MenuItem data-my-value="edit" onClick={(e) => handleClose(e)}>Edit</MenuItem>
-            <MenuItem data-my-value="open-close" onClick={(e) => handleClose(e)}>{data.status ? "Close" : "Open"}</MenuItem>
-            <MenuItem data-my-value="delete" onClick={(e) => handleClose(e)}>Delete</MenuItem>
-        </Menu>
+        {data && (
+          <Menu id="basic-menu" anchorEl={anchorEl} open={open} onClose={handleClose} MenuListProps={{ 'aria-labelledby': 'basic-button', }}>
+              <MenuItem data-my-value="edit" onClick={(e) => handleClose(e)}>Edit</MenuItem>
+              <MenuItem data-my-value="statusChange" onClick={(e) => handleClose(e)}>{data.status ? "Close" : "Open"}</MenuItem>
+              <MenuItem data-my-value="delete" onClick={(e) => handleClose(e)}>Delete</MenuItem>
+          </Menu>
+        )}
+        {dialogOpen && (
+          <ModifyDialog dialogOpen={dialogOpen} setDialogOpen={setDialogOpen} handleSubmit={handleSubmit} textContent={textContent}/>
+        )}
+        
     </>
   )
 }
